chore: 폴더 구조 및 userStore 파일 분리
